it works

+++ models/BusinessHours.model.js
@@ -9,7 +9,7 @@ class Business extends Model {
 					primaryKey: true, // Set id as the primary key
 					autoIncrement: true, // Enable auto-increment for the primary key
 				},
-				store_id: Sequelize.INTEGER,
+				store_id: Sequelize.BIGINT,
 				day_of_week: Sequelize.INTEGER,
 				start_time_local: Sequelize.TIME,
 				end_time_local: Sequelize.TIME,
